requirements.txt
absl-py==2.2.2
adata==2.9.4
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==12.1.0
baostock==0.8.9
beautifulsoup4==4.13.5
braceexpand==0.1.7
certifi==2026.2.25
charset-normalizer==2.1.1
click==8.3.1
cmake==3.25.0
coloredlogs==15.0.1
commandlines==0.4.1
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cupy-cuda11x==13.4.1
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.25.2
filetype==1.2.0
fire==0.7.0
flatbuffers==25.2.10
freetype-py==2.5.1
frozenlist==1.6.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
hjson==3.1.0
hsh==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
intel-cmplr-lib-ur==2025.3.3
intel-openmp==2025.3.3
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiui==0.2.16
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
lightning-utilities==0.15.3
lit==15.0.7
llvmlite==0.44.0
lpips==0.1.4
Markdown==3.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mkl==2025.3.1
mkl-service==2.4.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-ml-py==12.575.51
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
objprint==0.3.0
omegaconf==2.3.0
onemkl-license==2025.3.1
onnxruntime==1.22.0
open_clip_torch==3.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==12.2.0
PIMS==0.7
platformdirs==4.3.8
plyfile==1.1.3
pooch==1.8.2
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
py-mini-racer==0.6.0
pycparser==3.0
pydantic==2.11.5
pydantic_core==2.33.2
pyDeprecate==0.3.1
pyglet==2.1.13
Pygments==2.19.2
PyMatting==1.1.14
pynvml==12.0.0
PyOpenGL==3.1.0
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.6.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
questionary==2.1.1
referencing==0.36.2
regex==2024.11.6
requests==2.28.1
rich==14.2.0
roma==1.5.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
slicerator==1.1.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tbb==2022.3.1
tcmlib==1.4.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.5.10
timm==1.0.26
tokenizers==0.19.1
torch==2.1.0+cu118
torchmetrics==1.9.0
torchvision==0.16.0+cu118
tqdm==4.67.1
transformers==4.43.4
trimesh==4.11.4
triton==2.1.0
typeguard==4.4.4
typer==0.24.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.2
umf==1.0.3
urllib3==1.26.13
varname==0.14.0
waitress==3.0.2
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
xformers==0.0.22.post7+cu118
yarl==1.20.0
yunchang==0.6.3.post1
